// packages/front/modules/frontend/src/model/RevenueCount.model.ts
export interface IRevenueCount {
  symbol: string;
  todayRevenue: number;
  yesterdayRevenue: number;
  sevenDaysRevenue: number;
  thirtyOneDaysRevenue: number;
  todayCount: number;
  yesterdayCount: number;
  sevenDaysCount: number;
  thirtyOneDaysCount: number;
}

export default IRevenueCount;